Move helpers to torch.distributed.utils
Move these helpers to torch.distributed.utils as they will be consumed
by DDP for native mixed precision implementation
ghstack-source-id: 179693214

Differential Revision: [D42674745](https://our.internmc.facebook.com/intern/diff/D42674745/)

[ghstack-poisoned]
