+++ torch/distributed/utils.py
@@ -1,4 +1,6 @@
-from typing import Any, Dict, List, Tuple
+from typing import Any, Dict, List, Tuple, Callable, Union, Set, OrderedDict
+import dataclasses
+import traceback
 
 import torch
 import torch.distributed as dist
@@ -94,6 +96,92 @@ def _recursive_to(inputs, target_gpu, use_side_stream_for_tensor_copies):
         to_map = None  # type: ignore[assignment]
     return res
 
+def _p_assert(cond: Any, s: str, raise_assertion_error: bool = True) -> None:
+    """This is used as an alternate to ``assert`` when in the backward context
+    to print the error message ``s`` since otherwise, it is swallowed."""
+    if not cond:
+        print(s)
+        traceback.print_stack()
+        if raise_assertion_error:
+            raise AssertionError(s)
+
+def _alloc_storage(tensor: torch.Tensor, size: torch.Size) -> bool:
+    """
+    Allocate storage for ``tensor`` with the given size.
+
+    Returns:
+        bool: ``True`` if this method allocated storage and ``False`` if the
+        storage was already allocated.
+    """
+    with torch.no_grad():
+        already_allocated = tensor._typed_storage()._size() == size.numel()
+        if not already_allocated:
+            tensor_storage_size = tensor._typed_storage()._size()
+            _p_assert(
+                tensor_storage_size == 0,
+                f"Tensor storage should have been resized to be 0 but got {tensor_storage_size}",
+            )
+            tensor._typed_storage()._resize_(size.numel())
+        return not already_allocated
+
+
+def _free_storage(tensor: torch.Tensor) -> bool:
+    """
+    Frees the underlying storage of ``tensor``.
+
+    Returns:
+        bool: ``True`` if the method freed the storage and ``False`` if the
+        storage was already freed.
+    """
+    with torch.no_grad():
+        already_freed = tensor._typed_storage()._size() == 0
+        if not already_freed:
+            _p_assert(
+                tensor.storage_offset() == 0,
+                "Freeing a tensor's storage is unsafe when it is not the sole occupant\n"
+                f"storage offset: {tensor.storage_offset()}\n"
+                f"storage size: {tensor._typed_storage()._size()}\n"
+                f"tensor shape: {tensor.shape}",
+            )
+            tensor._typed_storage()._resize_(0)
+        return not already_freed
+
+def _apply_to_tensors(
+    fn: Callable,
+    container: Union[torch.Tensor, Dict, List, Tuple, Set, OrderedDict, PackedSequence],
+) -> Any:
+    """Recursively apply to all tensor in different kinds of container types."""
+
+    def apply(
+        x: Union[torch.Tensor, Dict, List, Tuple, Set, OrderedDict, PackedSequence]
+    ) -> Any:
+        if torch.is_tensor(x):
+            return fn(x)
+        elif hasattr(x, "__dataclass_fields__"):
+            dc = dataclasses.replace(x)
+            for f in dataclasses.fields(dc):
+                name = f.name
+                setattr(dc, name, apply(getattr(dc, name)))
+            return dc
+        elif isinstance(x, OrderedDict):
+            od = x.__class__()
+            for key, value in x.items():
+                od[key] = apply(value)
+            return od
+        elif isinstance(x, PackedSequence):
+            apply(x.data)
+            return x
+        elif isinstance(x, dict):
+            return {key: apply(value) for key, value in x.items()}
+        elif _is_namedtuple(x):
+            res = (apply(el) for el in x)
+            return type(x)(*res)
+        elif isinstance(x, (list, tuple, set)):
+            return type(x)(apply(el) for el in x)
+        else:
+            return x
+
+    return apply(container)
 
 def _to_kwargs(inputs, kwargs, device_id, use_side_stream_for_tensor_copies):
     inputs = (
